requirements.txt
torch
pandas
numpy
pandas-ta
joblib
PyYAML
gym
ray[rllib]==2.2.0
requests
python-binance
aiohttp
kafka-python
redis
pytz
tqdm
uvicorn
fastapi
pydantic
httpx